Extracted PreviewValidation to separate class
