+++ Zoop/ZoopUtility.cs
@@ -20,6 +20,8 @@ public static class ZoopUtility
 
   private static readonly ZoopSession Session = new();
   private static readonly ZoopPreviewFactory PreviewFactory = new(Session);
+  private static readonly ZoopPreviewValidator PreviewValidator =
+    new(ResolveBuildIndex, GetConstructableForBuildIndex, allowPlacementUpdate => AllowPlacementUpdate = allowPlacementUpdate);
 
   public static int PreviewCount => Session.PreviewCount;
   public static bool HasError { get => Session.HasError; private set => Session.HasError = value; }
@@ -31,30 +33,6 @@ public static class ZoopUtility
   }
 
   private const float PositionToleranceSqr = 0.0001f;
-
-  private readonly struct ValidationCursorState
-  {
-    public ValidationCursorState(int originalBuildIndex, int validationBuildIndex, bool needsCursorSwap,
-      Vector3 originalCursorPosition, Vector3 originalCursorLocalPosition, Quaternion originalCursorRotation,
-      Quaternion originalCursorLocalRotation)
-    {
-      OriginalBuildIndex = originalBuildIndex;
-      ValidationBuildIndex = validationBuildIndex;
-      NeedsCursorSwap = needsCursorSwap;
-      OriginalCursorPosition = originalCursorPosition;
-      OriginalCursorLocalPosition = originalCursorLocalPosition;
-      OriginalCursorRotation = originalCursorRotation;
-      OriginalCursorLocalRotation = originalCursorLocalRotation;
-    }
-
-    public int OriginalBuildIndex { get; }
-    public int ValidationBuildIndex { get; }
-    public bool NeedsCursorSwap { get; }
-    public Vector3 OriginalCursorPosition { get; }
-    public Vector3 OriginalCursorLocalPosition { get; }
-    public Quaternion OriginalCursorRotation { get; }
-    public Quaternion OriginalCursorLocalRotation { get; }
-  }
 
   private static readonly FieldInfo UsePrimaryPositionField =
     typeof(InventoryManager).GetField("_usePrimaryPosition", BindingFlags.Instance | BindingFlags.NonPublic);
@@ -918,8 +896,7 @@ public static class ZoopUtility
   /// </summary>
   private static bool CanConstructSmallCell(InventoryManager inventoryManager, Structure structure, int structureIndex)
   {
-    return CanConstructWithValidationCursor(inventoryManager, structure, structureIndex,
-      validationCursor => ValidateSmallCellCursor(inventoryManager, validationCursor));
+    return PreviewValidator.CanConstructSmallCell(inventoryManager, structure, structureIndex);
   }
 
   /// <summary>
@@ -927,189 +904,7 @@ public static class ZoopUtility
   /// </summary>
   private static bool CanConstructBigCell(InventoryManager inventoryManager, Structure structure, int structureIndex)
   {
-    return CanConstructWithValidationCursor(inventoryManager, structure, structureIndex, ValidateLargeGridCursor);
-  }
-
-  /// <summary>
-  /// Validates a zoop preview by temporarily retargeting the live construction cursor.
-  /// </summary>
-  private static bool CanConstructWithValidationCursor(InventoryManager inventoryManager, Structure structure,
-    int structureIndex, Func<Structure, bool> validator)
-  {
-    var originalCursor = InventoryManager.ConstructionCursor;
-    if (originalCursor == null)
-    {
-      return false;
-    }
-
-    // Capture the currently selected cursor before we temporarily retarget it to the zoop preview piece.
-    if (!TryGetValidationTarget(inventoryManager, structure, structureIndex, originalCursor,
-          out var validationConstructable, out var cursorState))
-    {
-      return false;
-    }
-
-    AllowPlacementUpdate = true;
-    try
-    {
-      // Rebuild the live cursor so the game validates this preview as if the player were placing it directly.
-      var validationCursor = PrepareValidationCursor(inventoryManager, structure, validationConstructable,
-        cursorState);
-      if (validationCursor == null)
-      {
-        return false;
-      }
-
-      return validator(validationCursor);
-    }
-    catch
-    {
-      return false;
-    }
-    finally
-    {
-      RestoreValidationCursor(inventoryManager, cursorState);
-    }
-  }
-
-  /// <summary>
-  /// Resolves the constructable and cursor snapshot needed for temporary small-grid validation.
-  /// </summary>
-  private static bool TryGetValidationTarget(InventoryManager inventoryManager, Structure structure,
-    int structureIndex, Structure originalCursor, out Structure validationConstructable,
-    out ValidationCursorState cursorState)
-  {
-    var originalBuildIndex = inventoryManager.ConstructionPanel.BuildIndex;
-    var validationBuildIndex = ResolveBuildIndex(inventoryManager, structure, structureIndex);
-    validationConstructable = GetConstructableForBuildIndex(inventoryManager, validationBuildIndex);
-    if (validationConstructable == null)
-    {
-      cursorState = default;
-      return false;
-    }
-
-    var needsCursorSwap = originalBuildIndex != validationBuildIndex ||
-                          originalCursor.PrefabName != validationConstructable.PrefabName;
-    cursorState = new ValidationCursorState(
-      originalBuildIndex,
-      validationBuildIndex,
-      needsCursorSwap,
-      originalCursor.ThingTransformPosition,
-      originalCursor.ThingTransformLocalPosition,
-      originalCursor.ThingTransformRotation,
-      originalCursor.ThingTransformLocalRotation);
-    return true;
-  }
-
-  /// <summary>
-  /// Prepares the live construction cursor so the game can validate one zoop preview piece.
-  /// </summary>
-  private static Structure PrepareValidationCursor(InventoryManager inventoryManager, Structure structure,
-    Structure validationConstructable, ValidationCursorState cursorState)
-  {
-    if (cursorState.NeedsCursorSwap)
-    {
-      inventoryManager.ConstructionPanel.BuildIndex = cursorState.ValidationBuildIndex;
-      InventoryManager.UpdatePlacement(validationConstructable);
-    }
-
-    var validationCursor = InventoryManager.ConstructionCursor;
-    if (validationCursor == null)
-    {
-      return null;
-    }
-
-    ApplyStructurePlacementState(validationCursor, structure.ThingTransformPosition, structure.ThingTransformLocalPosition,
-      structure.ThingTransformRotation, structure.ThingTransformLocalRotation);
-    validationCursor.CheckBounds();
-    validationCursor.RebuildGridState();
-    return validationCursor;
-  }
-
-  /// <summary>
-  /// Applies the same native placement checks the game uses for the live preview cursor.
-  /// </summary>
-  private static bool ValidateSmallCellCursor(InventoryManager inventoryManager, Structure validationCursor)
-  {
-    var canConstruct = validationCursor.CanConstruct().CanConstruct;
-    if (!canConstruct || InventoryManager.IsAuthoringMode || validationCursor is not IGridMergeable mergeable)
-    {
-      return canConstruct;
-    }
-
-    // Non-authoring previews do an extra merge/tool check for cables, pipes, and other grid-mergeables.
-    var activeConstructor = InventoryManager.Parent.Slots[inventoryManager.ActiveHand.SlotId].Get() as MultiConstructor;
-    var inactiveHandOccupant = InventoryManager.Parent.Slots[inventoryManager.InactiveHand.SlotId].Get() as Item;
-    return mergeable.CanReplace(activeConstructor, inactiveHandOccupant).CanConstruct;
-  }
-
-  /// <summary>
-  /// Checks whether a large-grid preview should be considered valid by the live cursor.
-  /// </summary>
-  private static bool ValidateLargeGridCursor(Structure validationCursor)
-  {
-    if (validationCursor is Wall)
-    {
-      var canMount = validationCursor.CanMountOnWall();
-      if (!canMount)
-      {
-        return false;
-      }
-    }
-
-    return validationCursor.CanConstruct().CanConstruct;
-  }
-
-  /// <summary>
-  /// Restores the live construction cursor after a temporary small-grid validation check.
-  /// </summary>
-  private static void RestoreValidationCursor(InventoryManager inventoryManager,
-    ValidationCursorState cursorState)
-  {
-    try
-    {
-      if (cursorState.NeedsCursorSwap)
-      {
-        inventoryManager.ConstructionPanel.BuildIndex = cursorState.OriginalBuildIndex;
-        var originalConstructable = GetConstructableForBuildIndex(inventoryManager, cursorState.OriginalBuildIndex);
-        if (originalConstructable != null)
-        {
-          InventoryManager.UpdatePlacement(originalConstructable);
-        }
-      }
-
-      var restoredCursor = InventoryManager.ConstructionCursor;
-      if (!cursorState.NeedsCursorSwap && restoredCursor != null)
-      {
-        ApplyStructurePlacementState(restoredCursor, cursorState.OriginalCursorPosition,
-          cursorState.OriginalCursorLocalPosition, cursorState.OriginalCursorRotation,
-          cursorState.OriginalCursorLocalRotation);
-      }
-
-      if (restoredCursor != null)
-      {
-        restoredCursor.gameObject.SetActive(false);
-      }
-    }
-    finally
-    {
-      AllowPlacementUpdate = false;
-    }
-  }
-
-  /// <summary>
-  /// Applies world and local placement state to a structure.
-  /// </summary>
-  private static void ApplyStructurePlacementState(Structure structure, Vector3 position, Vector3 localPosition,
-    Quaternion rotation, Quaternion localRotation)
-  {
-    structure.ThingTransformPosition = position;
-    structure.Position = position;
-    structure.transform.position = position;
-    structure.ThingTransformLocalPosition = localPosition;
-    structure.ThingTransformLocalRotation = localRotation;
-    structure.ThingTransformRotation = rotation;
-    structure.transform.rotation = rotation;
+    return PreviewValidator.CanConstructBigCell(inventoryManager, structure, structureIndex);
   }
 
   #endregion
